Move components out of AssistantBuilder file
