+++ front/lib/api/assistant/templates.ts
@@ -11,7 +11,7 @@ import {
   removeNulls,
 } from "@dust-tt/types";
 
-import type { BuilderFlow } from "@app/components/assistant_builder/AssistantBuilder";
+import type { BuilderFlow } from "@app/components/assistant_builder/types";
 import { TemplateResource } from "@app/lib/resources/template_resource";
 
 export async function generateMockAgentConfigurationFromTemplate(
