+++ front/pages/w/[wId]/builder/assistants/create.tsx
@@ -20,10 +20,10 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import { createRef, useEffect, useRef, useState } from "react";
 
-import type { BuilderFlow } from "@app/components/assistant_builder/AssistantBuilder";
-import { BUILDER_FLOWS } from "@app/components/assistant_builder/AssistantBuilder";
 import { AssistantTemplateModal } from "@app/components/assistant_builder/AssistantTemplateModal";
 import { TemplateGrid } from "@app/components/assistant_builder/TemplateGrid";
+import type { BuilderFlow } from "@app/components/assistant_builder/types";
+import { BUILDER_FLOWS } from "@app/components/assistant_builder/types";
 import AppLayout, { appLayoutBack } from "@app/components/sparkle/AppLayout";
 import { AppLayoutSimpleCloseTitle } from "@app/components/sparkle/AppLayoutTitle";
 import config from "@app/lib/api/config";
